conver HeatMap2 to HeatMap
